fix: add separate fcmRestNotificationService
